+++ slash/conf.py
@@ -62,7 +62,6 @@ config = Config({
         "localtime": False // Doc("Use local time for logging. If False, will use UTC"),
         "unittest_mode": False // Doc("Used during unit testing. Emit all logs to stderr as well as the log files"),
         "unified_session_log": False // Doc("Make the session log file contain all logs, including from tests"),
-        "disable_introspection": False // Doc("Don't output introspection when assertion fails"),
     },
     "run": {
         "dump_variation": False // Doc("Output the full variation structure before each test is run (mainly used for internal debugging)"),
@@ -75,6 +74,7 @@ config = Config({
         "session_state_path": "~/.slash/last_session" // Doc("Where to keep last session serialized data"),
         "project_customization_file_path": "./.slashrc",
         "user_customization_file_path": "~/.slash/slashrc",
+        "message_assertion_introspection": True // Doc("When False, failing assertions which have messages attached will not emit introspection info"),
     },
     "parallel": {
         "num_workers": 0 // Doc("Parallel execution") // Cmdline(arg='--parallel', metavar="NUM_WORKERS"),